Concentrating core crypto functionality
